Port JS `describe()` as Rust modules
